requirements.txt
anyio==3.6.1
asgiref==3.5.2
beanie==1.11.0
click==8.1.3
dnspython==2.2.1
fastapi==0.85.1
fastapi-pagination==0.10.0
h11==0.14.0
idna==3.4
motor==2.5.1
multidict==6.0.2
numpy==1.23.4
pandas==1.5.1
pydantic==1.10.2
pymongo==3.12.3
python-dateutil==2.8.2
python-decouple==3.6
pytz==2022.5
six==1.16.0
sniffio==1.3.0
starlette==0.20.4
toml==0.10.2
typing_extensions==4.4.0
uvicorn==0.17.6
yarl==1.8.1
